+++ appinfo/routes.php
@@ -7,5 +7,6 @@ return [
 		['name' => 'config#setConfig', 'url' => '/config', 'verb' => 'PUT'],
 		['name' => 'config#test', 'url' => '/config/test', 'verb' => 'GET'],
 		['name' => 'seal#seal', 'url' => '/seal/{fileId}', 'verb' => 'POST'],
+		['name' => 'seal#info', 'url' => '/seal/{fileId}', 'verb' => 'GET'],
 	],
 ];